backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Infrastructure Parsing
python-hcl2==4.3.2
pyyaml==6.0.1
boto3==1.34.34

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML
openai>=1.50.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# AWS S3 Compatible (for Spaces)
boto3==1.34.34

# PDF Processing
reportlab==4.0.9
pypdf2==3.0.1

# GitHub Integration
PyGithub==2.1.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
